requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
click==8.0.1
colorama==0.4.4
dj-database-url==1.0.0
Django==3.2.5
django-appconf==1.0.4
django-autoslug==1.9.8
django-bootstrap4==3.0.1
django-cities==0.6
django-cities-light==3.8.1
django-cors-headers==3.13.0
django-filter==2.4.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-model-utils==4.1.1
django-phonenumber-field==5.2.0
djangorestframework==3.13.1
Flask==2.0.1
GDAL @ file:///C:/gambby/GDAL-3.3.1-cp39-cp39-win_amd64.whl
geos==0.2.3
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.1
phonenumbers==8.12.27
pilkit==2.0
Pillow==8.3.1
progressbar2==3.53.1
psycopg2==2.9.1
pygeos==0.10.1
python-utils==2.5.6
pytz==2021.1
Shapely==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tqdm==4.61.2
Unidecode==1.2.0
Werkzeug==2.0.1
whitenoise==6.2.0
